[naming] host code now has its own file
